removing table
